audiopolicy: sort devices based on type, id and addr
Sort devices added to policy manager based on device type.
For devices of same type sort based on id, which places the
latest devices in the front of the list.
For devices of same type and no id, sort based on address.
For devices of same type, no id and address, fallback to default
sort using pointer address.

CRs-Fixed: 2504401
Change-Id: I25e2266d3d9f3df72ef86e7cbdc9389ead04a169
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>